Move RPM upload and publish tests
Move RPM upload and publish tests from module
`pulp_smash.tests.rpm.api_v2.test_sync_publish to module
`pulp_smash.tests.rpm.api_v2.test_upload_publish`. This has several
benefits:

* The `test_sync_publish` module is now as advertised: it only contains
  tests that sync and publish repositories.
* The task of uploading and publishing RPM repositories is a broad one,
  as a variety of content types can come in to play. They include RPM
  modules, erratum, and `comps.xml` files. There's now a clear home for
  these tests.
