+++ docs/api/pulp_smash.tests.rpm.api_v2.test_upload_publish.rst
@@ -1,0 +1,6 @@
+`pulp_smash.tests.rpm.api_v2.test_upload_publish`
+=================================================
+
+Location: :doc:`/index` → :doc:`/api` → :doc:`/api/pulp_smash.tests.rpm.api_v2.test_upload_publish`
+
+.. automodule:: pulp_smash.tests.rpm.api_v2.test_upload_publish
